requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.2.0
python-dotenv>=1.0.1
pydantic>=2.7.0
rich>=13.7.0
sqlalchemy>=2.0.0
